requirements.txt
# Telegram Bot Framework
aiogram==3.15.0
aiohttp==3.11.11

# Database
python-dotenv==1.0.1

# HTTP Requests
requests==2.32.5
httpx==0.28.1

# Scheduler
apscheduler==3.11.1

# Utilities
python-dateutil==2.9.0.post0

# Security & Encryption
cryptography==42.0.5

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Optional: Supabase support
# supabase==2.10.0
